restructure extension shared state
