+++ package.json
@@ -5,9 +5,9 @@
   "author": "Damian Tarnawski <gthetarnav@gmail.com>",
   "contributors": [],
   "scripts": {
-    "dev:ext": "turbo run dev --filter=solid-devtools-extension",
+    "dev:ext": "pnpm -dir packages/extension run dev",
     "dev:ext-deps": "turbo run dev --filter=@solid-devtools/shared --filter=@solid-devtools/ui",
-    "dev:sandbox": "turbo run dev --filter=solid-devtools-sandbox-playground",
+    "dev:sandbox": "pnpm -dir playgrounds/sandbox run dev",
     "build": "turbo run build --filter=!./playgrounds/*",
     "test": "turbo run test --filter=!./playgrounds/*",
     "typecheck": "turbo run typecheck --filter=!./playgrounds/*",
